chore: rename remote -> trace
