+++ consensus/state.go
@@ -24,7 +24,7 @@ import (
 	"github.com/tendermint/tendermint/libs/service"
 	tmsync "github.com/tendermint/tendermint/libs/sync"
 	"github.com/tendermint/tendermint/p2p"
-	"github.com/tendermint/tendermint/pkg/remote"
+	"github.com/tendermint/tendermint/pkg/trace"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	sm "github.com/tendermint/tendermint/state"
 	"github.com/tendermint/tendermint/types"
@@ -142,7 +142,7 @@ type State struct {
 	// for reporting metrics
 	metrics *Metrics
 
-	eventCollector *remote.Client
+	eventCollector *trace.Client
 }
 
 // StateOption sets an optional parameter on the State.
@@ -173,7 +173,7 @@ func NewState(
 		evpool:           evpool,
 		evsw:             tmevents.NewEventSwitch(),
 		metrics:          NopMetrics(),
-		eventCollector:   &remote.Client{},
+		eventCollector:   &trace.Client{},
 	}
 
 	// set function defaults (may be overwritten before calling Start)
@@ -216,7 +216,7 @@ func StateMetrics(metrics *Metrics) StateOption {
 }
 
 // SetEventCollector sets the remote event collector.
-func SetEventCollector(ec *remote.Client) StateOption {
+func SetEventCollector(ec *trace.Client) StateOption {
 	return func(cs *State) { cs.eventCollector = ec }
 }
