+++ pkg/trace/README.md
