+++ pkg/trace/client.go
@@ -1,4 +1,4 @@
-package remote
+package trace
 
 import (
 	"context"
