+++ test/e2e/runner/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/tendermint/tendermint/libs/log"
-	"github.com/tendermint/tendermint/pkg/remote"
+	"github.com/tendermint/tendermint/pkg/trace"
 	e2e "github.com/tendermint/tendermint/test/e2e/pkg"
 	"github.com/tendermint/tendermint/test/e2e/pkg/infra"
 	"github.com/tendermint/tendermint/test/e2e/pkg/infra/docker"
@@ -79,11 +79,11 @@ func NewCLI() *CLI {
 				return fmt.Errorf("unknown infrastructure type '%s'", inft)
 			}
 
-			iurl, err := cmd.Flags().GetString(remote.FlagInfluxDBURL)
+			iurl, err := cmd.Flags().GetString(trace.FlagInfluxDBURL)
 			if err != nil {
 				return err
 			}
-			itoken, err := cmd.Flags().GetString(remote.FlagInfluxDBToken)
+			itoken, err := cmd.Flags().GetString(trace.FlagInfluxDBToken)
 			if err != nil {
 				return err
 			}
@@ -174,9 +174,9 @@ func NewCLI() *CLI {
 
 	cli.root.PersistentFlags().StringP("infrastructure-data", "", "", "path to the json file containing the infrastructure data. Only used if the 'infrastructure-type' is set to a value other than 'docker'")
 
-	cli.root.PersistentFlags().String(remote.FlagInfluxDBURL, "", remote.FlagInfluxDBURLDescription)
+	cli.root.PersistentFlags().String(trace.FlagInfluxDBURL, "", trace.FlagInfluxDBURLDescription)
 
-	cli.root.PersistentFlags().String(remote.FlagInfluxDBToken, "", remote.FlagInfluxDBTokenDescription)
+	cli.root.PersistentFlags().String(trace.FlagInfluxDBToken, "", trace.FlagInfluxDBTokenDescription)
 
 	cli.root.Flags().BoolVarP(&cli.preserve, "preserve", "p", false,
 		"Preserves the running of the test net after tests are completed")
